% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Field.R
\name{ForeignKey}
\alias{ForeignKey}
\title{Reference a field from another table.}
\usage{
ForeignKey(relation, field = NULL)
}
\arguments{
\item{relation}{An \code{koral_relation} object.}

\item{field}{The name of the referenced field. Required in case \code{relation} has a
composite key. If \code{NULL}, defined as the PK at \code{relation}.}
}
\description{
Reference a field from another table.
}
