% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sql-ops.R
\name{get_all}
\alias{get_all}
\title{Get all records from a given relation.}
\usage{
get_all(relation, ..., .limit = NULL)
}
\arguments{
\item{relation}{An \code{koral_relation} object.}

\item{...}{Key-value pairs representing filters applied to the table.}

\item{.limit}{Integer. If not \code{NULL}, includes a \code{LIMIT} clause.}
}
\value{
List of lists: inner lists contain the retrieved records.
}
\description{
Get all records from a given relation.
}
