% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sql-ops.R
\name{get_one}
\alias{get_one}
\title{Get a single record satisfying key-value filters in the dots.}
\usage{
get_one(relation, ...)
}
\value{
A \code{list} representing the record.
}
\description{
Get a single record satisfying key-value filters in the dots.
}
