% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sql-ops.R
\name{insert}
\alias{insert}
\title{Insert an object into a table.}
\usage{
insert(relation, x)
}
\arguments{
\item{relation}{An \code{koral_relation} object.}

\item{x}{A \code{list} representing the record to be parsed and saved.}
}
\value{
A \code{list} representing the saved record, including its generated fields.
}
\description{
Insert an object into a table.
}
