% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sql-ops.R
\name{create_table}
\alias{create_table}
\alias{drop_table}
\title{Create a table in the database for the given \code{relation}.}
\usage{
create_table(relation, drop_if_exists = FALSE)

drop_table(relation)
}
\arguments{
\item{relation}{An \code{koral_relation} object.}

\item{drop_if_exists}{Logical. If \code{TRUE}, drops the table with the same name (in case
it exists) before creating the given one.}
}
\value{
Logical indicating whether the table was created.

Logical indicating whether the table was dropped.
}
\description{
Create a table in the database for the given \code{relation}.

Drops the table \code{attr(relation, "table")} in the database.
}
