% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Relation.R
\name{Relation}
\alias{Relation}
\title{Declare the model for a relation of items.}
\usage{
Relation(table, ...)
}
\arguments{
\item{table}{The name of the table.}

\item{...}{Sequence of \code{koral_field} declarations (or lists of those, to be unnested).}
}
\value{
An \code{koral_relation} object. It's a named list of \code{koral_field}s with an
\code{"table"} attribute.
}
\description{
Declare the model for a relation of items.
}
